+++ featuregate/flags.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package featuregate // import "go.opentelemetry.io/collector/service/featuregate"
+package featuregate // import "go.opentelemetry.io/collector/featuregate"
 
 import (
 	"flag"